Unsigned integer compatibility added.

+++ src/com/herestt/ro2/vdk/VDK1FileInfo.java
@@ -6,15 +6,15 @@ import java.util.Map;
 public class VDK1FileInfo extends VDKInnerDirectory{
 
 	private String version;							// VDK file's version.
-	private int unknown;							// Unknown.
-	private int fileCount;							// Amount of file in the VDK file.
-	private int folderCount;						// Amount of folder in the VDK file.
-	private int size;								// VDK file's size.	
-	private int fileListPartLength;					// Trailing file list's size.
-	private Map<Integer, String> filePathMap;		// Map holding the trailing file list. <File_Offset, File_Path>.
+	private long unknown;							// Unknown.
+	private long fileCount;							// Amount of file in the VDK file.
+	private long folderCount;						// Amount of folder in the VDK file.
+	private long size;								// VDK file's size.	
+	private long fileListPartLength;					// Trailing file list's size.
+	private Map<Long, String> filePathMap;		// Map holding the trailing file list. <File_Offset, File_Path>.
 	
-	public VDK1FileInfo(String version, int unknown, int fileCount,	int folderCount, int size, 
-			int fileListPartSize) {
+	public VDK1FileInfo(String version, long unknown, long fileCount, long folderCount, long size, 
+			long fileListPartSize) {
 		
 		super();
 		this.version = version;
@@ -36,51 +36,51 @@ public class VDK1FileInfo extends VDKInnerDirectory{
 		this.version = version;
 	}
 
-	public int getUnknown() {
+	public long getUnknown() {
 		return unknown;
 	}
 
-	public void setUnknown(int unknown) {
+	public void setUnknown(long unknown) {
 		this.unknown = unknown;
 	}
 
-	public int getFileCount() {
+	public long getFileCount() {
 		return fileCount;
 	}
 
-	public void setFileCount(int fileCount) {
+	public void setFileCount(long fileCount) {
 		this.fileCount = fileCount;
 	}
 
-	public int getFolderCount() {
+	public long getFolderCount() {
 		return folderCount;
 	}
 
-	public void setFolderCount(int folderCount) {
+	public void setFolderCount(long folderCount) {
 		this.folderCount = folderCount;
 	}
 
-	public int getSize() {
+	public long getSize() {
 		return size;
 	}
 
-	public void setSize(int size) {
+	public void setSize(long size) {
 		this.size = size;
 	}
 
-	public int getFileListPartLength() {
+	public long getFileListPartLength() {
 		return fileListPartLength;
 	}
 
-	public void setFileListPartLength(int fileListPartLength) {
+	public void setFileListPartLength(long fileListPartLength) {
 		this.fileListPartLength = fileListPartLength;
 	}
 
-	public Map<Integer, String> getFilePathMap() {
+	public Map<Long, String> getFilePathMap() {
 		return filePathMap;
 	}
 
-	public void setFilePathMap(Map<Integer, String> filePathMap) {
+	public void setFilePathMap(Map<Long, String> filePathMap) {
 		this.filePathMap = filePathMap;
 	}
